Add files via upload
My First Webpage,
This is a simple HTML and CSS project for a "One Piece Fan Club" web page. It includes a styled header and footer, a short introduction about One Piece with a white background paragraph, and an image of the crew(crew.jpg). A translucent registration form is provided for fans to join the club by submitting their name, email, and favorite character. The page uses a background image (logoo.jpg) to enhance the design and create an engaging look for anime fans.